LPD-70013 Adapting test to move the addTemporaryIndex tests from old class to new class
