+++ modules/apps/portal/portal-dao-test/src/testIntegration/java/com/liferay/portal/dao/db/test/BaseDBProcessTest.java
@@ -7,7 +7,9 @@ package com.liferay.portal.dao.db.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.function.UnsafeConsumer;
+import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.dao.db.BaseDB;
 import com.liferay.portal.kernel.dao.db.BaseDBProcess;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBInspector;
@@ -16,6 +18,7 @@ import com.liferay.portal.kernel.dao.db.IndexMetadata;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
@@ -34,6 +37,7 @@ import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
+import java.util.concurrent.atomic.AtomicLong;
 
 import org.junit.After;
 import org.junit.AfterClass;
@@ -63,6 +67,9 @@ public class BaseDBProcessTest extends BaseDBProcess {
 		_dbInspector = new DBInspector(_connection);
 
 		_db = DBManagerUtil.getDB();
+
+		_tempIndexCounter = ReflectionTestUtil.getFieldValue(
+			BaseDB.class, "_tempIndexCounter");
 	}
 
 	@AfterClass
@@ -89,6 +96,62 @@ public class BaseDBProcessTest extends BaseDBProcess {
 	@After
 	public void tearDown() throws Exception {
 		runSQL("DROP_TABLE_IF_EXISTS(" + _TABLE_NAME + ")");
+	}
+
+	@Test
+	public void testAddMultipleTemporaryIndex() throws Exception {
+		UpgradeProcess upgradeProcess = new UpgradeProcess() {
+
+			@Override
+			protected void doUpgrade() throws Exception {
+				String tempIndexName1 =
+					"IX_TEMP_" + (_tempIndexCounter.get() + 1);
+				String tempIndexName2 =
+					"IX_TEMP_" + (_tempIndexCounter.get() + 2);
+
+				DB db = DBManagerUtil.getDB();
+
+				try (SafeCloseable safeCloseable1 = db.addTemporaryIndex(
+						connection, _TABLE_NAME, false, "typeVarchar");
+					SafeCloseable safeCloseable2 = db.addTemporaryIndex(
+						connection, _TABLE_NAME, false, "id", "typeVarchar")) {
+
+					Assert.assertTrue(hasIndex(_TABLE_NAME, tempIndexName1));
+					Assert.assertTrue(hasIndex(_TABLE_NAME, tempIndexName2));
+				}
+
+				Assert.assertFalse(hasIndex(_TABLE_NAME, tempIndexName1));
+				Assert.assertFalse(hasIndex(_TABLE_NAME, tempIndexName2));
+			}
+
+		};
+
+		upgradeProcess.upgrade();
+	}
+
+	@Test
+	public void testAddTemporaryIndex() throws Exception {
+		UpgradeProcess upgradeProcess = new UpgradeProcess() {
+
+			@Override
+			protected void doUpgrade() throws Exception {
+				String tempIndexName =
+					"IX_TEMP_" + (_tempIndexCounter.get() + 1);
+
+				DB db = DBManagerUtil.getDB();
+
+				try (SafeCloseable safeCloseable = db.addTemporaryIndex(
+						connection, _TABLE_NAME, false, "typeVarchar")) {
+
+					Assert.assertTrue(hasIndex(_TABLE_NAME, tempIndexName));
+				}
+
+				Assert.assertFalse(hasIndex(_TABLE_NAME, tempIndexName));
+			}
+
+		};
+
+		upgradeProcess.upgrade();
 	}
 
 	@Test
@@ -663,5 +726,6 @@ public class BaseDBProcessTest extends BaseDBProcess {
 	private static Connection _connection;
 	private static DB _db;
 	private static DBInspector _dbInspector;
+	private static AtomicLong _tempIndexCounter;
 
 }
